Resolve "Update README gifs with new terrain data"

Closes #98

See merge request fireline/simfire!85

+++ run_game.py
@@ -88,27 +88,30 @@ def main():
 
     if cfg.simulation.record:
         out_path = os.curdir + "/simulation.gif"
-        game.frames[0].save(out_path, save_all=True, duration=100, loop=0)
-
-    fig = fire_manager.draw_spread_graph(game.screen)
-    if cfg.simulation.headless:
-        save_path = os.curdir + "fire_spread_graph.png"
-        print(
-            "Game is running in a headless state. Saving fire spread "
-            f"graph to {save_path}"
+        game.frames[0].save(
+            out_path, save_all=True, duration=100, loop=0, append_images=game.frames[1:]
         )
-        fig.savefig(save_path)
-    else:
-        if "DISPLAY" in os.environ:
-            print(
-                "Game is running in a non-headless state. Displaying fire spread "
-                f'graph on DISPLAY {os.environ["DISPLAY"]}'
-            )
-            import matplotlib.pyplot as plt
 
-            plt.show()
-            while plt.fignum_exists(fig.number):
-                continue
+    if cfg.simulation.draw_spread_graph:
+        fig = fire_manager.draw_spread_graph(game.screen)
+        if cfg.simulation.headless:
+            save_path = os.curdir + "fire_spread_graph.png"
+            print(
+                "Game is running in a headless state. Saving fire spread "
+                f"graph to {save_path}"
+            )
+            fig.savefig(save_path)
+        else:
+            if "DISPLAY" in os.environ:
+                print(
+                    "Game is running in a non-headless state. Displaying fire spread "
+                    f'graph on DISPLAY {os.environ["DISPLAY"]}'
+                )
+                import matplotlib.pyplot as plt
+
+                plt.show()
+                while plt.fignum_exists(fig.number):
+                    continue
 
 
 if __name__ == "__main__":
